feat: basic implementation of application
This can read the extensions directly from `vscode`, parse them, download them, hash them, and finally output a valid Nix expression to the `stdout`. I've taken the liberty to include a `vscode` package in the devshell that comes with a bunch of extensions, populated through this very tool actually. Note that it produces the newer SRI-style hashes that Nix uses for derivation, instead of the legacy sha256.