feat: functional implementation without argument support
This is a huge mileston. For the first time since the beginning of this
project, we've successfully managed to produce 100% reproducible
derivations through the generated `extensions.nix` file. This partially
implements the ideas of #2, but is not yet complete. A major refactor
needs to happen first, followed by more fine-tuning of the queries, then
finally support for CLI arguments.