feat: major refactor of processing, divide into smaller modules
Massive changes have been made, so let's jot them down one by one.

The `api` module contains all of the code relevant to serializing and
deserializing queries and response payloads from communicating with the
vscode api. It matches the exact structure, but has many fields missing,
as it's not intended to be a re-implementation.

The `models` module contains the structure model of a vscode extension,
which is used to serialize and deserialize the structure to and from
json via `serde_json`. For now the deserializing isn't implemented
because I removed the ability to read from a file, but I'm working on
bringing that back. The serializing is useful to write the final Nix
expression to `stdout`.

The `nix` module contains some helper functions to interact with Nix.
There is a hash function that converts `&[u8]` to a sha256 hash in sri
format. There is another function that takes a serializable structure
and converts it into json, then turns it into a Nix expression string.

The main processing in the application has undergone significant
changes. Firstly, it'll process all the extensions by their 'name'
(which comprises the publisher and actual extension name, separated by a
dot). It runs this through the api to query information about the
extension. The information we recieve contains the name, publisher,
version, any target platforms if they are necessary, and a uri from
which we can download that exact revision of the extension. This is
super useful because it saves us a lot of tedious one-by-one processing
time by looking at the `package.json` and `extension.vsixmanifest`. We
use the uri to download the extension, and hash it. We then continue to
produce the final nix expression, pretty much exactly like before.