feat: implement arguments to read nix file and output to path (or stdout)
Very small improvement to the CLI to accept an option `-f|--file` argument, that will read the extensions from a nix file that was previously generated through this app, or from `code --list-extensions` if not given. There is also an optional `-o|--out` which will write the final expression to the specified path, or stdout if not given.